Major reorganization of the package structure moving forms to individual packages as well as moving button panels as well with matching display panels.
Rework of the Display panel and made static so that Actions can directly
make calls to the display. The Display was broken into upper and lower
sections with scroll panes and the upper holds a visual panel while the
lower holds a table that we replace the data model to change what is
displayed.

Actions were also updated to begin to display the correct panels.