Change the way the primary display panel displays the sub-panels.
The original problem was that the display panel was created but never
passed to the outside actions and tried to display gui components
outside of the event thread. This was changed so that we call the
primary gui for the only display panel and then change it's internal
panel/table model so that all gui panels work inside of the event
thread.